Adds Anime X (BETA) to header and sidebar
Adds a link to the Anime X (BETA) page in both the header and sidebar for easier access.

Also moves the search input above the spotlight section and improves the look of the anime spotlight section on the Anime X index page.

Adds a recommendation section to the Anime X show page.